Add Button functionality
The code in the src\pages\api\addComic.js partially works to my knowledge. Apparently the first time I ran the add button everything got added to the database. This meant the reason for the errors I was getting was because I did not add a catch into the create method that would allow the database to check if the person existed in the current table and link it to the new comic entry. I implemented this check so far on only the writer method and will need to implement on the rest of the contributor methods.